download.pl: use curl in preference to wget
Because wget doesn't know how to do Negotiate authentication with a proxy and curl does, use curl if it's present. The user is expected to have a ~/.curlrc that sets the options necessary for any proxy authentication. A ~/.curlrc is completely optional however and curl will work in exactly the same manner as wget without one.

Signed-off-by: Brian J. Murrell <brian@interlinx.bc.ca>